+++ src/legato/api/mkBuild.ts
@@ -4,7 +4,9 @@ import { pathExists } from 'fs-extra';
 import * as path from 'path';
 import { Command } from '../../commons/identifiers';
 import { CommandRegister } from '../../commons/manager';
-import { LegatoFileExtension, LegatoManager } from './core';
+import { LegatoManager } from './core';
+import { LegatoFileExtension } from './files';
+import { getEvalExpression, ModelElement } from '../../commons/model';
 
 /**
  * Build tools
@@ -46,24 +48,40 @@ export class MkBuildManager extends CommandRegister {
     }
 
     /**
+     * @param param the name of the param to fill
+     * @param elt the model element to use a the value of the given param
+     * @returns the ouple of parameters if the value exist and an empty array if not
+     */
+    private async toParameters(param: string, elt: ModelElement<string | undefined>): Promise<[string, string] | []> {
+        if (await elt.get()) {
+            return [param, getEvalExpression(elt)];
+        }
+        return [];
+    }
+
+    /**
      * @param mktool the current mktool
      * @returns the corresponding build command
      */
     public async getBuildCommand(): Promise<string | undefined> {
         const mktool = await this.getMkTool();
         if (mktool) {
-            let command = `${mktool} ${this.legatoManager.defFile.getEvalExpression()} -s components -t ${this.legatoManager.legatoTarget.getEvalExpression()}`;
+            let command = [
+                mktool, getEvalExpression(this.legatoManager.defFile),
+                "-s", "components",
+                "-t", getEvalExpression(this.legatoManager.target)
+            ];
 
             // object dir arg complete if not empty
-            command = await this.legatoManager.legatoObjectDir.get() ? command.concat(" -w ", this.legatoManager.legatoObjectDir.getEvalExpression()) : command;
+            command.push(...await this.toParameters("-w", this.legatoManager.objectDir));
 
             // output dir arg complete if not empty
-            command = await this.legatoManager.legatoOutputDir.get() ? command.concat(" -o ", this.legatoManager.legatoOutputDir.getEvalExpression()) : command;
+            command.push(...await this.toParameters("-o", this.legatoManager.outputDir));
 
             // debug directory
-            command = await this.legatoManager.legatoDebugDir.get() ? command.concat(" -d ", this.legatoManager.legatoDebugDir.getEvalExpression()) : command;
+            command.push(...await this.toParameters("-d", this.legatoManager.debugDir));
 
-            return command;
+            return command.join(" ");
         }
 
         // unavailable build command when no def file is present
@@ -77,7 +95,7 @@ export class MkBuildManager extends CommandRegister {
     public async getBuildAndInstallCommand(): Promise<string | undefined> {
         const buildCommand = await this.getBuildCommand();
         if (buildCommand) {
-            return `${buildCommand} && update ${this.legatoManager.legatoUpdateFile.getEvalExpression()}`;
+            return `${buildCommand} && update ${getEvalExpression(this.legatoManager.updateFile)}`;
         }
         return undefined;
     }
@@ -87,14 +105,14 @@ export class MkBuildManager extends CommandRegister {
      */
     public async getCleanCommand() {
         const artifacts: Array<string> = [];
-        if (await this.legatoManager.legatoDebugDir.get()) {
-            artifacts.push(this.legatoManager.legatoDebugDir.getEvalExpression());
+        if (await this.legatoManager.debugDir.get()) {
+            artifacts.push(getEvalExpression(this.legatoManager.debugDir));
         }
-        if (await this.legatoManager.legatoObjectDir.get()) {
-            artifacts.push(this.legatoManager.legatoObjectDir.getEvalExpression());
+        if (await this.legatoManager.objectDir.get()) {
+            artifacts.push(getEvalExpression(this.legatoManager.objectDir));
         }
-        if (await this.legatoManager.legatoUpdateFile.get()) {
-            artifacts.push(this.legatoManager.legatoUpdateFile.getEvalExpression());
+        if (await this.legatoManager.updateFile.get()) {
+            artifacts.push(getEvalExpression(this.legatoManager.updateFile));
         }
 
         return artifacts.length > 0 ? "rm -rf ".concat(...artifacts.map(value => value.concat(" "))) : undefined;
@@ -105,11 +123,10 @@ export class MkBuildManager extends CommandRegister {
      */
     public async generateImage() {
         //if the .update file exists, the generate image command is provided
-        const updateFile = await this.legatoManager.legatoUpdateFile.getResolvedPath();
+        const updateFile = await this.legatoManager.updateFile.get();
         if (updateFile && await pathExists(updateFile)) {
-            return `systoimg ${this.legatoManager.legatoTarget.getEvalExpression()} ${this.legatoManager.legatoUpdateFile.getEvalExpression()} ${this.legatoManager.legatoObjectDir.getEvalExpression()}/image`;
+            return `systoimg ${getEvalExpression(this.legatoManager.target)} ${getEvalExpression(this.legatoManager.updateFile)} ${getEvalExpression(this.legatoManager.objectDir)}/image`;
         }
         return undefined;
     }
-
 }